Separate logic for transfer task execution and queue management
