+++ service/history/transferQueueActiveTaskExecutor_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 Uber Technologies, Inc.
+// Copyright (c) 2020 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -54,7 +54,7 @@ import (
 )
 
 type (
-	transferQueueActiveProcessorSuite struct {
+	transferQueueActiveTaskExecutorSuite struct {
 		suite.Suite
 		*require.Assertions
 
@@ -77,37 +77,37 @@ type (
 		mockArchiverProvider        *provider.MockArchiverProvider
 		mockParentClosePolicyClient *parentclosepolicy.ClientMock
 
-		logger                       log.Logger
-		domainID                     string
-		domainName                   string
-		domainEntry                  *cache.DomainCacheEntry
-		targetDomainID               string
-		targetDomainName             string
-		targetDomainEntry            *cache.DomainCacheEntry
-		childDomainID                string
-		childDomainName              string
-		childDomainEntry             *cache.DomainCacheEntry
-		version                      int64
-		now                          time.Time
-		timeSource                   *clock.EventTimeSource
-		transferQueueActiveProcessor *transferQueueActiveProcessorImpl
+		logger                          log.Logger
+		domainID                        string
+		domainName                      string
+		domainEntry                     *cache.DomainCacheEntry
+		targetDomainID                  string
+		targetDomainName                string
+		targetDomainEntry               *cache.DomainCacheEntry
+		childDomainID                   string
+		childDomainName                 string
+		childDomainEntry                *cache.DomainCacheEntry
+		version                         int64
+		now                             time.Time
+		timeSource                      *clock.EventTimeSource
+		transferQueueActiveTaskExecutor *transferQueueActiveTaskExecutor
 	}
 )
 
-func TestTransferQueueActiveProcessorSuite(t *testing.T) {
-	s := new(transferQueueActiveProcessorSuite)
+func TestTransferQueueActiveTaskExecutorSuite(t *testing.T) {
+	s := new(transferQueueActiveTaskExecutorSuite)
 	suite.Run(t, s)
 }
 
-func (s *transferQueueActiveProcessorSuite) SetupSuite() {
+func (s *transferQueueActiveTaskExecutorSuite) SetupSuite() {
 
 }
 
-func (s *transferQueueActiveProcessorSuite) TearDownSuite() {
+func (s *transferQueueActiveTaskExecutorSuite) TearDownSuite() {
 
 }
 
-func (s *transferQueueActiveProcessorSuite) SetupTest() {
+func (s *transferQueueActiveTaskExecutorSuite) SetupTest() {
 	s.Assertions = require.New(s.T())
 
 	s.domainID = testDomainID
@@ -131,6 +131,7 @@ func (s *transferQueueActiveProcessorSuite) SetupTest() {
 	s.mockReplicationProcessor.EXPECT().notifyNewTask().AnyTimes()
 	s.mockTimerProcessor.EXPECT().NotifyNewTimers(gomock.Any(), gomock.Any()).AnyTimes()
 
+	config := NewDynamicConfigForTest()
 	s.mockShard = newTestShardContext(
 		s.controller,
 		&persistence.ShardInfo{
@@ -138,11 +139,12 @@ func (s *transferQueueActiveProcessorSuite) SetupTest() {
 			RangeID:          1,
 			TransferAckLevel: 0,
 		},
-		NewDynamicConfigForTest(),
+		config,
 	)
 	s.mockShard.eventsCache = newEventsCache(s.mockShard)
 	s.mockShard.resource.TimeSource = s.timeSource
 
+	s.mockParentClosePolicyClient = &parentclosepolicy.ClientMock{}
 	s.mockArchivalClient = &warchiver.ClientMock{}
 	s.mockMatchingClient = s.mockShard.resource.MatchingClient
 	s.mockHistoryClient = s.mockShard.resource.HistoryClient
@@ -188,30 +190,24 @@ func (s *transferQueueActiveProcessorSuite) SetupTest() {
 	s.mockShard.SetEngine(h)
 
 	s.mockQueueAckMgr = &MockQueueAckMgr{}
-	s.transferQueueActiveProcessor = newTransferQueueActiveProcessor(
+	s.transferQueueActiveTaskExecutor = newTransferQueueActiveTaskExecutor(
 		s.mockShard,
 		h,
-		s.mockVisibilityMgr,
-		s.mockMatchingClient,
-		s.mockHistoryClient,
-		newTaskAllocator(s.mockShard),
 		s.logger,
-	)
-	s.transferQueueActiveProcessor.queueAckMgr = s.mockQueueAckMgr
-	s.transferQueueActiveProcessor.queueProcessorBase.ackMgr = s.mockQueueAckMgr
-
-	s.mockParentClosePolicyClient = &parentclosepolicy.ClientMock{}
-	s.transferQueueActiveProcessor.parentClosePolicyClient = s.mockParentClosePolicyClient
+		s.mockShard.GetMetricsClient(),
+		config,
+	).(*transferQueueActiveTaskExecutor)
+	s.transferQueueActiveTaskExecutor.parentClosePolicyClient = s.mockParentClosePolicyClient
 }
 
-func (s *transferQueueActiveProcessorSuite) TearDownTest() {
+func (s *transferQueueActiveTaskExecutorSuite) TearDownTest() {
 	s.controller.Finish()
 	s.mockShard.Finish(s.T())
 	s.mockQueueAckMgr.AssertExpectations(s.T())
 	s.mockArchivalClient.AssertExpectations(s.T())
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessActivityTask_Success() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessActivityTask_Success() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -261,11 +257,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessActivityTask_Success() {
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockMatchingClient.EXPECT().AddActivityTask(gomock.Any(), s.createAddActivityTaskRequest(transferTask, ai)).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessActivityTask_Duplication() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessActivityTask_Duplication() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -318,11 +314,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessActivityTask_Duplication(
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_FirstDecision() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessDecisionTask_FirstDecision() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -364,11 +360,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_FirstDecisio
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockMatchingClient.EXPECT().AddDecisionTask(gomock.Any(), s.createAddDecisionTaskRequest(transferTask, mutableState)).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_NonFirstDecision() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessDecisionTask_NonFirstDecision() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -417,11 +413,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_NonFirstDeci
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockMatchingClient.EXPECT().AddDecisionTask(gomock.Any(), s.createAddDecisionTaskRequest(transferTask, mutableState)).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_Sticky_NonFirstDecision() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessDecisionTask_Sticky_NonFirstDecision() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -476,11 +472,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_Sticky_NonFi
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockMatchingClient.EXPECT().AddDecisionTask(gomock.Any(), s.createAddDecisionTaskRequest(transferTask, mutableState)).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_DecisionNotSticky_MutableStateSticky() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessDecisionTask_DecisionNotSticky_MutableStateSticky() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -535,11 +531,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_DecisionNotS
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockMatchingClient.EXPECT().AddDecisionTask(gomock.Any(), s.createAddDecisionTaskRequest(transferTask, mutableState)).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_Duplication() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessDecisionTask_Duplication() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -583,11 +579,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessDecisionTask_Duplication(
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_HasParent() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCloseExecution_HasParent() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -656,11 +652,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_HasParent(
 	s.mockVisibilityMgr.On("RecordWorkflowExecutionClosed", mock.Anything).Return(nil).Once()
 	s.mockArchivalMetadata.On("GetVisibilityConfig").Return(archiver.NewDisabledArchvialConfig())
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCloseExecution_NoParent() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -709,11 +705,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent()
 	s.mockArchivalMetadata.On("GetVisibilityConfig").Return(archiver.NewArchivalConfig("enabled", dc.GetStringPropertyFn("enabled"), dc.GetBoolPropertyFn(true), "disabled", "random URI"))
 	s.mockArchivalClient.On("Archive", mock.Anything, mock.Anything).Return(nil, nil).Once()
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_HasFewChildren() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCloseExecution_NoParent_HasFewChildren() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -843,11 +839,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_H
 	s.mockHistoryClient.EXPECT().RequestCancelWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 	s.mockHistoryClient.EXPECT().TerminateWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_HasManyChildren() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCloseExecution_NoParent_HasManyChildren() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -934,11 +930,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_H
 	s.mockArchivalMetadata.On("GetVisibilityConfig").Return(archiver.NewDisabledArchvialConfig())
 	s.mockParentClosePolicyClient.On("SendParentClosePolicyRequest", mock.Anything).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_HasManyAbandonedChildren() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCloseExecution_NoParent_HasManyAbandonedChildren() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1024,11 +1020,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCloseExecution_NoParent_H
 	s.mockVisibilityMgr.On("RecordWorkflowExecutionClosed", mock.Anything).Return(nil).Once()
 	s.mockArchivalMetadata.On("GetVisibilityConfig").Return(archiver.NewDisabledArchvialConfig())
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Success() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCancelExecution_Success() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1086,11 +1082,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Success()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&p.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &p.MutableStateUpdateSessionStats{}}, nil).Once()
 	s.mockClusterMetadata.EXPECT().ClusterNameForFailoverVersion(s.version).Return(cluster.TestCurrentClusterName).AnyTimes()
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Failure() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCancelExecution_Failure() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1148,11 +1144,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Failure()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&p.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &p.MutableStateUpdateSessionStats{}}, nil).Once()
 	s.mockClusterMetadata.EXPECT().ClusterNameForFailoverVersion(s.version).Return(cluster.TestCurrentClusterName).AnyTimes()
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Duplication() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessCancelExecution_Duplication() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1208,11 +1204,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessCancelExecution_Duplicati
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Success() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessSignalExecution_Success() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1283,11 +1279,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Success()
 		RequestId: common.StringPtr(si.SignalRequestID),
 	}).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Failure() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessSignalExecution_Failure() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1349,11 +1345,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Failure()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&p.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &p.MutableStateUpdateSessionStats{}}, nil).Once()
 	s.mockClusterMetadata.EXPECT().ClusterNameForFailoverVersion(s.version).Return(cluster.TestCurrentClusterName).AnyTimes()
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Duplication() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessSignalExecution_Duplication() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1413,11 +1409,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessSignalExecution_Duplicati
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Success() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessStartChildExecution_Success() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1491,11 +1487,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Succe
 		IsFirstDecision: common.BoolPtr(true),
 	}).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Failure() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessStartChildExecution_Failure() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1570,11 +1566,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Failu
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&p.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &p.MutableStateUpdateSessionStats{}}, nil).Once()
 	s.mockClusterMetadata.EXPECT().ClusterNameForFailoverVersion(s.version).Return(cluster.TestCurrentClusterName).AnyTimes()
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Success_Dup() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessStartChildExecution_Success_Dup() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1651,11 +1647,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Succe
 		IsFirstDecision: common.BoolPtr(true),
 	}).Return(nil).Times(1)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Duplication() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessStartChildExecution_Duplication() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1730,11 +1726,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessStartChildExecution_Dupli
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessRecordWorkflowStartedTask() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessRecordWorkflowStartedTask() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1781,11 +1777,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessRecordWorkflowStartedTask
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockVisibilityMgr.On("RecordWorkflowExecutionStarted", s.createRecordWorkflowExecutionStartedRequest(s.domainName, event, transferTask, mutableState, backoffSeconds)).Once().Return(nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestProcessUpsertWorkflowSearchAttributes() {
+func (s *transferQueueActiveTaskExecutorSuite) TestProcessUpsertWorkflowSearchAttributes() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1828,11 +1824,11 @@ func (s *transferQueueActiveProcessorSuite) TestProcessUpsertWorkflowSearchAttri
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 	s.mockVisibilityMgr.On("UpsertWorkflowExecution", s.createUpsertWorkflowSearchAttributesRequest(s.domainName, event, transferTask, mutableState)).Once().Return(nil)
 
-	_, err = s.transferQueueActiveProcessor.process(newTaskInfo(nil, transferTask, s.logger))
+	err = s.transferQueueActiveTaskExecutor.execute(newTaskInfo(nil, transferTask, s.logger))
 	s.Nil(err)
 }
 
-func (s *transferQueueActiveProcessorSuite) TestCopySearchAttributes() {
+func (s *transferQueueActiveTaskExecutorSuite) TestCopySearchAttributes() {
 	var input map[string][]byte
 	s.Nil(copySearchAttributes(input))
 
@@ -1847,7 +1843,7 @@ func (s *transferQueueActiveProcessorSuite) TestCopySearchAttributes() {
 	s.Equal(byte('1'), val[0])
 }
 
-func (s *transferQueueActiveProcessorSuite) createAddActivityTaskRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createAddActivityTaskRequest(
 	task *persistence.TransferTaskInfo,
 	ai *persistence.ActivityInfo,
 ) *matching.AddActivityTaskRequest {
@@ -1868,7 +1864,7 @@ func (s *transferQueueActiveProcessorSuite) createAddActivityTaskRequest(
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createAddDecisionTaskRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createAddDecisionTaskRequest(
 	task *persistence.TransferTaskInfo,
 	mutableState mutableState,
 ) *matching.AddDecisionTaskRequest {
@@ -1894,7 +1890,7 @@ func (s *transferQueueActiveProcessorSuite) createAddDecisionTaskRequest(
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createRecordWorkflowExecutionStartedRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createRecordWorkflowExecutionStartedRequest(
 	domainName string,
 	startEvent *workflow.HistoryEvent,
 	task *persistence.TransferTaskInfo,
@@ -1920,7 +1916,7 @@ func (s *transferQueueActiveProcessorSuite) createRecordWorkflowExecutionStarted
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createRequestCancelWorkflowExecutionRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createRequestCancelWorkflowExecutionRequest(
 	targetDomainName string,
 	task *persistence.TransferTaskInfo,
 	rci *persistence.RequestCancelInfo,
@@ -1950,7 +1946,7 @@ func (s *transferQueueActiveProcessorSuite) createRequestCancelWorkflowExecution
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createSignalWorkflowExecutionRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createSignalWorkflowExecutionRequest(
 	targetDomainName string,
 	task *persistence.TransferTaskInfo,
 	si *persistence.SignalInfo,
@@ -1981,7 +1977,7 @@ func (s *transferQueueActiveProcessorSuite) createSignalWorkflowExecutionRequest
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createChildWorkflowExecutionRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createChildWorkflowExecutionRequest(
 	domainName string,
 	childDomainName string,
 	task *persistence.TransferTaskInfo,
@@ -2023,7 +2019,7 @@ func (s *transferQueueActiveProcessorSuite) createChildWorkflowExecutionRequest(
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createUpsertWorkflowSearchAttributesRequest(
+func (s *transferQueueActiveTaskExecutorSuite) createUpsertWorkflowSearchAttributesRequest(
 	domainName string,
 	startEvent *workflow.HistoryEvent,
 	task *persistence.TransferTaskInfo,
@@ -2047,7 +2043,7 @@ func (s *transferQueueActiveProcessorSuite) createUpsertWorkflowSearchAttributes
 	}
 }
 
-func (s *transferQueueActiveProcessorSuite) createPersistenceMutableState(
+func (s *transferQueueActiveTaskExecutorSuite) createPersistenceMutableState(
 	ms mutableState,
 	lastEventID int64,
 	lastEventVersion int64,
