+++ service/history/transferQueueStandbyProcessor.go
@@ -21,11 +21,7 @@
 package history
 
 import (
-	"time"
-
-	workflow "github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/client/matching"
-	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
@@ -39,18 +35,13 @@ type (
 		*queueProcessorBase
 		queueAckMgr
 
-		clusterName         string
-		shard               ShardContext
-		config              *Config
-		historyService      *historyEngineImpl
-		options             *QueueProcessorOptions
-		executionManager    persistence.ExecutionManager
-		cache               *historyCache
-		transferTaskFilter  taskFilter
-		logger              log.Logger
-		metricsClient       metrics.Client
-		historyRereplicator xdc.HistoryRereplicator
-		nDCHistoryResender  xdc.NDCHistoryResender
+		clusterName        string
+		shard              ShardContext
+		config             *Config
+		transferTaskFilter taskFilter
+		logger             log.Logger
+		metricsClient      metrics.Client
+		taskExecutor       queueTaskExecutor
 	}
 )
 
@@ -101,27 +92,27 @@ func newTransferQueueStandbyProcessor(
 		clusterName:        clusterName,
 		shard:              shard,
 		config:             shard.GetConfig(),
-		historyService:     historyService,
-		options:            options,
-		executionManager:   shard.GetExecutionManager(),
-		cache:              historyService.historyCache,
 		transferTaskFilter: transferTaskFilter,
 		logger:             logger,
 		metricsClient:      historyService.metricsClient,
+		taskExecutor: newTransferQueueStandbyTaskExecutor(
+			shard,
+			historyService,
+			historyRereplicator,
+			nDCHistoryResender,
+			logger,
+			historyService.metricsClient,
+			clusterName,
+			config,
+		),
 		transferQueueProcessorBase: newTransferQueueProcessorBase(
 			shard,
-			config,
-			historyService,
 			options,
-			visibilityMgr,
-			matchingClient,
 			maxReadAckLevel,
 			updateClusterAckLevel,
 			transferQueueShutdown,
 			logger,
 		),
-		historyRereplicator: historyRereplicator,
-		nDCHistoryResender:  nDCHistoryResender,
 	}
 
 	queueAckMgr := newQueueAckMgr(shard, options, processor, shard.GetTransferClusterAckLevel(clusterName), logger)
@@ -150,574 +141,7 @@ func (t *transferQueueStandbyProcessorImpl) complete(
 func (t *transferQueueStandbyProcessorImpl) process(
 	taskInfo *taskInfo,
 ) (int, error) {
-
-	task, ok := taskInfo.task.(*persistence.TransferTaskInfo)
-	if !ok {
-		return metrics.TransferStandbyQueueProcessorScope, errUnexpectedQueueTask
-	}
-
-	var err error
-	switch task.TaskType {
-	case persistence.TransferTaskTypeActivityTask:
-		if taskInfo.shouldProcessTask {
-			err = t.processActivityTask(taskInfo)
-		}
-		return metrics.TransferStandbyTaskActivityScope, err
-
-	case persistence.TransferTaskTypeDecisionTask:
-		if taskInfo.shouldProcessTask {
-			err = t.processDecisionTask(taskInfo)
-		}
-		return metrics.TransferStandbyTaskDecisionScope, err
-
-	case persistence.TransferTaskTypeCloseExecution:
-		// guarantee the processing of workflow execution close
-		err = t.processCloseExecution(taskInfo)
-		return metrics.TransferStandbyTaskCloseExecutionScope, err
-
-	case persistence.TransferTaskTypeCancelExecution:
-		if taskInfo.shouldProcessTask {
-			err = t.processCancelExecution(taskInfo)
-		}
-		return metrics.TransferStandbyTaskCancelExecutionScope, err
-
-	case persistence.TransferTaskTypeSignalExecution:
-		if taskInfo.shouldProcessTask {
-			err = t.processSignalExecution(taskInfo)
-		}
-		return metrics.TransferStandbyTaskSignalExecutionScope, err
-
-	case persistence.TransferTaskTypeStartChildExecution:
-		if taskInfo.shouldProcessTask {
-			err = t.processStartChildExecution(taskInfo)
-		}
-		return metrics.TransferStandbyTaskStartChildExecutionScope, err
-
-	case persistence.TransferTaskTypeRecordWorkflowStarted:
-		if taskInfo.shouldProcessTask {
-			err = t.processRecordWorkflowStarted(taskInfo)
-		}
-		return metrics.TransferStandbyTaskRecordWorkflowStartedScope, err
-
-	case persistence.TransferTaskTypeResetWorkflow:
-		// no reset needed for standby
-		return metrics.TransferStandbyTaskResetWorkflowScope, err
-
-	case persistence.TransferTaskTypeUpsertWorkflowSearchAttributes:
-		if taskInfo.shouldProcessTask {
-			err = t.processUpsertWorkflowSearchAttributes(taskInfo)
-		}
-		return metrics.TransferStandbyTaskUpsertWorkflowSearchAttributesScope, err
-
-	default:
-		return metrics.TransferStandbyQueueProcessorScope, errUnknownTransferTask
-	}
-}
-
-func (t *transferQueueStandbyProcessorImpl) processActivityTask(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		activityInfo, ok := mutableState.GetActivityInfo(transferTask.ScheduleID)
-		if !ok {
-			return nil, nil
-		}
-
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, activityInfo.Version, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		if activityInfo.StartedID == common.EmptyEventID {
-			return newPushActivityToMatchingInfo(
-				activityInfo.ScheduleToStartTimeout,
-			), nil
-		}
-
-		return nil, nil
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.pushActivity,
-			t.pushActivity,
-		),
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processDecisionTask(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		decisionInfo, ok := mutableState.GetDecisionInfo(transferTask.ScheduleID)
-		if !ok {
-			return nil, nil
-		}
-
-		executionInfo := mutableState.GetExecutionInfo()
-		workflowTimeout := executionInfo.WorkflowTimeout
-		decisionTimeout := common.MinInt32(workflowTimeout, common.MaxTaskTimeout)
-
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, decisionInfo.Version, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		if decisionInfo.StartedID == common.EmptyEventID {
-			return newPushDecisionToMatchingInfo(
-				decisionTimeout,
-				workflow.TaskList{Name: &transferTask.TaskList},
-			), nil
-		}
-
-		return nil, nil
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.pushDecision,
-			t.pushDecision,
-		),
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processCloseExecution(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := true
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		if mutableState.IsWorkflowExecutionRunning() {
-			// this can happen if workflow is reset.
-			return nil, nil
-		}
-
-		completionEvent, err := mutableState.GetCompletionEvent()
-		if err != nil {
-			return nil, err
-		}
-		wfCloseTime := completionEvent.GetTimestamp()
-
-		executionInfo := mutableState.GetExecutionInfo()
-		workflowTypeName := executionInfo.WorkflowTypeName
-		workflowCloseTimestamp := wfCloseTime
-		workflowCloseStatus := persistence.ToThriftWorkflowExecutionCloseStatus(executionInfo.CloseStatus)
-		workflowHistoryLength := mutableState.GetNextEventID() - 1
-		startEvent, err := mutableState.GetStartEvent()
-		if err != nil {
-			return nil, err
-		}
-		workflowStartTimestamp := startEvent.GetTimestamp()
-		workflowExecutionTimestamp := getWorkflowExecutionTimestamp(mutableState, startEvent)
-		visibilityMemo := getWorkflowMemo(executionInfo.Memo)
-		searchAttr := executionInfo.SearchAttributes
-
-		lastWriteVersion, err := mutableState.GetLastWriteVersion()
-		if err != nil {
-			return nil, err
-		}
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, lastWriteVersion, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		// DO NOT REPLY TO PARENT
-		// since event replication should be done by active cluster
-		return nil, t.recordWorkflowClosed(
-			transferTask.DomainID,
-			transferTask.WorkflowID,
-			transferTask.RunID,
-			workflowTypeName,
-			workflowStartTimestamp,
-			workflowExecutionTimestamp.UnixNano(),
-			workflowCloseTimestamp,
-			workflowCloseStatus,
-			workflowHistoryLength,
-			transferTask.GetTaskID(),
-			visibilityMemo,
-			searchAttr,
-		)
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		standbyTaskPostActionNoOp,
-	) // no op post action, since the entire workflow is finished
-}
-
-func (t *transferQueueStandbyProcessorImpl) processCancelExecution(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		requestCancelInfo, ok := mutableState.GetRequestCancelInfo(transferTask.ScheduleID)
-		if !ok {
-			return nil, nil
-		}
-
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, requestCancelInfo.Version, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		return getHistoryResendInfo(mutableState)
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTransferTaskPostActionTaskDiscarded,
-		),
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processSignalExecution(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		signalInfo, ok := mutableState.GetSignalInfo(transferTask.ScheduleID)
-		if !ok {
-			return nil, nil
-		}
-
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, signalInfo.Version, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		return getHistoryResendInfo(mutableState)
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTransferTaskPostActionTaskDiscarded,
-		),
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processStartChildExecution(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-		transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-		childWorkflowInfo, ok := mutableState.GetChildExecutionInfo(transferTask.ScheduleID)
-		if !ok {
-			return nil, nil
-		}
-
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, childWorkflowInfo.Version, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return nil, err
-		}
-
-		if childWorkflowInfo.StartedID != common.EmptyEventID {
-			return nil, nil
-		}
-
-		return getHistoryResendInfo(mutableState)
-	}
-
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTransferTaskPostActionTaskDiscarded,
-		),
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processRecordWorkflowStarted(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-			transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-
-			return nil, t.processRecordWorkflowStartedOrUpsertHelper(transferTask, mutableState, true)
-		},
-		standbyTaskPostActionNoOp,
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processUpsertWorkflowSearchAttributes(
-	taskInfo *taskInfo,
-) error {
-
-	processTaskIfClosed := false
-	return t.processTransfer(
-		processTaskIfClosed,
-		taskInfo,
-		func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
-
-			transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-			return nil, t.processRecordWorkflowStartedOrUpsertHelper(transferTask, mutableState, false)
-		},
-		standbyTaskPostActionNoOp,
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) processRecordWorkflowStartedOrUpsertHelper(
-	transferTask *persistence.TransferTaskInfo,
-	mutableState mutableState,
-	isRecordStart bool,
-) error {
-
-	// verify task version for RecordWorkflowStarted.
-	// upsert doesn't require verifyTask, because it is just a sync of mutableState.
-	if isRecordStart {
-		startVersion, err := mutableState.GetStartVersion()
-		if err != nil {
-			return err
-		}
-		ok, err := verifyTaskVersion(t.shard, t.logger, transferTask.DomainID, startVersion, transferTask.Version, transferTask)
-		if err != nil || !ok {
-			return err
-		}
-	}
-
-	executionInfo := mutableState.GetExecutionInfo()
-	workflowTimeout := executionInfo.WorkflowTimeout
-	wfTypeName := executionInfo.WorkflowTypeName
-	startEvent, err := mutableState.GetStartEvent()
-	if err != nil {
-		return err
-	}
-	startTimestamp := startEvent.GetTimestamp()
-	executionTimestamp := getWorkflowExecutionTimestamp(mutableState, startEvent)
-	visibilityMemo := getWorkflowMemo(executionInfo.Memo)
-	searchAttr := copySearchAttributes(executionInfo.SearchAttributes)
-
-	if isRecordStart {
-		return t.recordWorkflowStarted(
-			transferTask.DomainID,
-			transferTask.WorkflowID,
-			transferTask.RunID,
-			wfTypeName,
-			startTimestamp,
-			executionTimestamp.UnixNano(),
-			workflowTimeout,
-			transferTask.GetTaskID(),
-			visibilityMemo,
-			searchAttr,
-		)
-	}
-	return t.upsertWorkflowExecution(
-		transferTask.DomainID,
-		transferTask.WorkflowID,
-		transferTask.RunID,
-		wfTypeName,
-		startTimestamp,
-		executionTimestamp.UnixNano(),
-		workflowTimeout,
-		transferTask.GetTaskID(),
-		visibilityMemo,
-		searchAttr,
-	)
-
-}
-
-func (t *transferQueueStandbyProcessorImpl) processTransfer(
-	processTaskIfClosed bool,
-	taskInfo *taskInfo,
-	actionFn standbyActionFn,
-	postActionFn standbyPostActionFn,
-) (retError error) {
-
-	transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.getDomainIDAndWorkflowExecution(transferTask),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if retError == ErrTaskRetry {
-			release(nil)
-		} else {
-			release(retError)
-		}
-	}()
-
-	mutableState, err := loadMutableStateForTransferTask(context, transferTask, t.metricsClient, t.logger)
-	if err != nil || mutableState == nil {
-		return err
-	}
-
-	if !mutableState.IsWorkflowExecutionRunning() && !processTaskIfClosed {
-		// workflow already finished, no need to process the timer
-		return nil
-	}
-
-	historyResendInfo, err := actionFn(context, mutableState)
-	if err != nil {
-		return err
-	}
-
-	release(nil)
-	return postActionFn(taskInfo, historyResendInfo, taskInfo.logger)
-}
-
-func (t *transferQueueStandbyProcessorImpl) pushActivity(
-	task *taskInfo,
-	postActionInfo interface{},
-	logger log.Logger,
-) error {
-
-	if postActionInfo == nil {
-		return nil
-	}
-
-	pushActivityInfo := postActionInfo.(*pushActivityToMatchingInfo)
-	timeout := common.MinInt32(pushActivityInfo.activityScheduleToStartTimeout, common.MaxTaskTimeout)
-	return t.transferQueueProcessorBase.pushActivity(
-		task.task.(*persistence.TransferTaskInfo),
-		timeout,
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) pushDecision(
-	task *taskInfo,
-	postActionInfo interface{},
-	logger log.Logger,
-) error {
-
-	if postActionInfo == nil {
-		return nil
-	}
-
-	pushDecisionInfo := postActionInfo.(*pushDecisionToMatchingInfo)
-	timeout := common.MinInt32(pushDecisionInfo.decisionScheduleToStartTimeout, common.MaxTaskTimeout)
-	return t.transferQueueProcessorBase.pushDecision(
-		task.task.(*persistence.TransferTaskInfo),
-		&pushDecisionInfo.tasklist,
-		timeout,
-	)
-}
-
-func (t *transferQueueStandbyProcessorImpl) fetchHistoryFromRemote(
-	taskInfo *taskInfo,
-	postActionInfo interface{},
-	log log.Logger,
-) error {
-
-	if postActionInfo == nil {
-		return nil
-	}
-
-	transferTask := taskInfo.task.(*persistence.TransferTaskInfo)
-	resendInfo := postActionInfo.(*historyResendInfo)
-
-	t.metricsClient.IncCounter(metrics.HistoryRereplicationByTransferTaskScope, metrics.CadenceClientRequests)
-	stopwatch := t.metricsClient.StartTimer(metrics.HistoryRereplicationByTransferTaskScope, metrics.CadenceClientLatency)
-	defer stopwatch.Stop()
-
-	var err error
-	if resendInfo.lastEventID != nil && resendInfo.lastEventVersion != nil {
-		err = t.nDCHistoryResender.SendSingleWorkflowHistory(
-			transferTask.DomainID,
-			transferTask.WorkflowID,
-			transferTask.RunID,
-			resendInfo.lastEventID,
-			resendInfo.lastEventVersion,
-			nil,
-			nil,
-		)
-	} else if resendInfo.nextEventID != nil {
-		err = t.historyRereplicator.SendMultiWorkflowHistory(
-			transferTask.DomainID,
-			transferTask.WorkflowID,
-			transferTask.RunID,
-			*resendInfo.nextEventID,
-			transferTask.RunID,
-			common.EndEventID, // use common.EndEventID since we do not know where is the end
-		)
-	} else {
-		err = &workflow.InternalServiceError{
-			Message: "transferQueueStandbyProcessor encounter empty historyResendInfo",
-		}
-	}
-
-	if err != nil {
-		t.logger.Error("Error re-replicating history from remote.",
-			tag.ShardID(t.shard.GetShardID()),
-			tag.WorkflowDomainID(transferTask.DomainID),
-			tag.WorkflowID(transferTask.WorkflowID),
-			tag.WorkflowRunID(transferTask.RunID),
-			tag.SourceCluster(t.clusterName))
-	}
-
-	// return error so task processing logic will retry
-	return ErrTaskRetry
-}
-
-func (t *transferQueueStandbyProcessorImpl) getCurrentTime() time.Time {
-	return t.shard.GetCurrentTime(t.clusterName)
+	// TODO: task metricScope should be determined when creating taskInfo
+	metricScope := t.getTransferTaskMetricsScope(taskInfo.task.GetTaskType(), false)
+	return metricScope, t.taskExecutor.execute(taskInfo)
 }
